pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Praveen"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytz (>=2025.2,<2026.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
django = "^5.2.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
